missed packages exported, documentation added, test improvement
