+++ src/interface/chains.ts
@@ -1,0 +1,2 @@
+
+export type ChainId = "fuji" | "avalanche"
