--- src/interfaces/index.ts
@@ -1,4 +1,0 @@
-// interfaces/index.ts
-export type { Metadata } from './Metadata';
-export type { BlockchainAction } from './BlockchainAction';
-
